VBoxManage: build fix: limit keyboardputscancode 'typing' rate. Simple commands to 'type' ASCII chars.
svn:sync-xref-src-repo-rev: r112317

+++ src/VBox/Frontends/VBoxManage/VBoxManageControlVM.cpp
@@ -1,4 +1,4 @@
-/* $Id: VBoxManageControlVM.cpp 64906 2016-12-16 13:01:20Z vitali.pelenjow@oracle.com $ */
+/* $Id: VBoxManageControlVM.cpp 64907 2016-12-16 13:05:46Z vitali.pelenjow@oracle.com $ */
 /** @file
  * VBoxManage - Implementation of the controlvm command.
  */
@@ -32,6 +32,7 @@
 #include <iprt/getopt.h>
 #include <iprt/stream.h>
 #include <iprt/string.h>
+#include <iprt/thread.h>
 #include <iprt/uuid.h>
 #include <iprt/file.h>
 #include <VBox/log.h>
@@ -228,7 +229,7 @@ static HRESULT keyboardPutScancodes(IKeyboard *pKeyboard, const std::list<LONG> 
 
 #if 1
     HRESULT rc = S_OK;
-    int i;
+    size_t i;
     for (i = 0; i < saScancodes.size(); ++i)
     {
         rc = pKeyboard->PutScancode(saScancodes[i]);
